requirements.txt
Django==1.8.2
Pillow==2.8.2
django-bootstrap3==6.1.0
django-bootstrap3-datetimepicker==2.2.3
django-stronghold==0.2.7
djangorestframework==3.1.3
isoweek==1.3.0
reportlab==3.2.0
